# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Fintech Tools API — tool directory, tool requests, usage events
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ToolStatus = typeof ToolStatus[keyof typeof ToolStatus];


export const ToolStatus = {
  live: 'live',
  coming_soon: 'coming_soon',
} as const;

export interface ToolSummary {
  slug: string;
  name: string;
  blurb: string;
  status: ToolStatus;
  sortOrder: number;
}

export interface ToolDetail {
  slug: string;
  name: string;
  blurb: string;
  status: ToolStatus;
  sortOrder: number;
  seoTitle: string;
  seoDescription: string;
  /** Markdown with "## How it works", "## Who it's for" and "## FAQ" sections; FAQ pairs are "### Question" headings followed by answer paragraphs. */
  pageCopy: string;
}

export interface ToolRequestCreate {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  requestText: string;
  /** @maxLength 1000 */
  currentCostText?: string;
  /** @maxLength 320 */
  email?: string;
}

export type UsageEventCreateEventType = typeof UsageEventCreateEventType[keyof typeof UsageEventCreateEventType];


export const UsageEventCreateEventType = {
  calculate: 'calculate',
  notify_me: 'notify_me',
} as const;

export type UsageEventCreatePayload = { [key: string]: unknown };

export interface UsageEventCreate {
  toolSlug: string;
  eventType: UsageEventCreateEventType;
  payload?: UsageEventCreatePayload;
}

export interface CreatedResponse {
  id: number;
}

export interface ApiError {
  message: string;
}