# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fintech Tools API — tool directory, tool requests, usage events
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tools/{slug}:
    get:
      operationId: getToolBySlug
      tags: [tools]
      summary: Get a tool by slug
      description: Returns full tool detail including SEO fields and page copy. 404 for unknown or hidden tools.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tool detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ToolStatus:
      type: string
      enum: [live, coming_soon]
...
    ToolDetail:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        blurb:
          type: string
        status:
          $ref: "#/components/schemas/ToolStatus"
        sortOrder:
          type: number
        seoTitle:
          type: string
        seoDescription:
          type: string
        pageCopy:
          type: string
          description: Markdown with "## How it works", "## Who it's for" and "## FAQ" sections; FAQ pairs are "### Question" headings followed by answer paragraphs.
      required: [slug, name, blurb, status, sortOrder, seoTitle, seoDescription, pageCopy]
...
    ApiError:
      type: object
      properties:
        message:
          type: string
      required: [message]