# lib/api-zod/src/generated/types/apiError.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Fintech Tools API — tool directory, tool requests, usage events
 * OpenAPI spec version: 0.1.0
 */

export interface ApiError {
  message: string;
}