# lib/api-zod/src/generated/types/toolDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Fintech Tools API — tool directory, tool requests, usage events
 * OpenAPI spec version: 0.1.0
 */
import type { ToolStatus } from './toolStatus';

export interface ToolDetail {
  slug: string;
  name: string;
  blurb: string;
  status: ToolStatus;
  sortOrder: number;
  seoTitle: string;
  seoDescription: string;
  /** Markdown with "## How it works", "## Who it's for" and "## FAQ" sections; FAQ pairs are "### Question" headings followed by answer paragraphs. */
  pageCopy: string;
}