# lib/api-zod/src/generated/types/toolStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Fintech Tools API — tool directory, tool requests, usage events
 * OpenAPI spec version: 0.1.0
 */

export type ToolStatus = typeof ToolStatus[keyof typeof ToolStatus];


export const ToolStatus = {
  live: 'live',
  coming_soon: 'coming_soon',
} as const;